backend/requirements.txt
Flask==3.1.1
flask-cors==6.0.1
Flask-Bcrypt==1.0.1
Flask-Dance==7.1.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
flask-talisman
Werkzeug==3.1.3
python-dotenv==1.1.1
requests==2.32.4
Pillow==10.0.0
numpy==1.24.3
supabase==2.20.0
SQLAlchemy==2.0.35
psycopg2-binary
PyJWT==2.10.1
bcrypt==4.3.0
groq==0.4.1
google-generativeai==0.8.3
gunicorn==20.1.0

# ML Dependencies for trained models
tensorflow-cpu==2.15.0
ultralytics==8.3.47
opencv-python-headless==4.8.1.78
torch==2.0.1
torchvision==0.15.2
transformers==4.30.2
sentence-transformers==2.2.2
h5py==3.9.0
huggingface-hub==0.16.4
